COMP: Fix -Wsign-compare warning in vtkMRMLApplicationLogic.cxx
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=27375
git-svn-id: http://svn.slicer.org/Slicer4/trunk@27375 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/Logic/vtkMRMLApplicationLogic.cxx
@@ -863,7 +863,7 @@ bool vtkMRMLApplicationLogic::SaveSceneToSlicerDataBundleDirectory(const char* s
           {
           storageNode->ResetFileNameList();
           std::vector< std::string > &originalFileNames = this->OriginalStorageNodeFileNames[storageNode];
-          for (int i = 0; i < originalFileNames.size(); i++)
+          for (size_t i = 0; i < originalFileNames.size(); i++)
             {
             if (i == 0)
               {
@@ -887,7 +887,7 @@ bool vtkMRMLApplicationLogic::SaveSceneToSlicerDataBundleDirectory(const char* s
 
         storageNode->ResetFileNameList();
         std::vector< std::string > &originalFileNames = this->OriginalStorageNodeFileNames[storageNode];
-        for (int i = 0; i < originalFileNames.size(); i++)
+        for (size_t i = 0; i < originalFileNames.size(); i++)
           {
           if (i == 0)
             {
